- CommandManager now uses navdata acks for configuration commands - Priority queue fixed (comparator) - Connect Configuration manager permanently - many small improvements

+++ YADrone/src/com/shigeodayo/ardrone/command/KeepAliveCommand.java
@@ -13,7 +13,7 @@ public class KeepAliveCommand extends ATCommand {
 	}
 
 	@Override
-	public int getPriority() {
+	public byte getPriority() {
 		return VERY_HIGH_PRIORITY;
 	}
 }
