+++ YADrone/src/com/shigeodayo/ardrone/command/QuitCommand.java
@@ -2,7 +2,7 @@ package com.shigeodayo.ardrone.command;
 
 public class QuitCommand extends DroneCommand {
 	@Override
-	public int getPriority() {
+	public byte getPriority() {
 		return MAX_PRIORITY;
 	}
 }
